+++ app/services/forest_liana/resources_getter.rb
@@ -4,7 +4,8 @@ module ForestLiana
       @resource = resource
       @params = params
       @count_needs_includes = false
-      @collection = get_collection(@resource.table_name)
+      @collection_name = ForestLiana.name_for(@resource)
+      @collection = get_collection(@collection_name)
       @field_names_requested = field_names_requested
       get_segment()
     end
@@ -79,8 +80,7 @@ module ForestLiana
     end
 
     def field_names_requested
-      collection_name = ForestLiana.name_for(@resource)
-      return nil unless @params[:fields] && @params[:fields][collection_name]
+      return nil unless @params[:fields] && @params[:fields][@collection_name]
 
       associations_for_query = []
 
@@ -100,7 +100,7 @@ module ForestLiana
         associations_for_query << @params[:sort].split('.').first.to_sym
       end
 
-      field_names = @params[:fields][collection_name].split(',').map { |name| name.to_sym }
+      field_names = @params[:fields][@collection_name].split(',').map { |name| name.to_sym }
       field_names | associations_for_query
     end