+++ lib/forest_liana.rb
@@ -21,6 +21,8 @@ module ForestLiana
   mattr_accessor :included_models
   mattr_accessor :user_class_name
   mattr_accessor :names_overriden
+  # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
+  mattr_accessor :names_old_overriden
 
   self.apimap = []
   self.allowed_users = []
@@ -30,14 +32,22 @@ module ForestLiana
   self.user_class_name = nil
   self.names_overriden = {}
 
+  # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
+  self.names_old_overriden = {}
+
   def self.schema_for_resource resource
     self.apimap.find do |collection|
       SchemaUtils.find_model_from_collection_name(collection.name)
-        .try(:table_name) == resource.table_name
+        .try(:name) == resource.name
     end
   end
 
   def self.name_for(model)
-    self.names_overriden[model] || model.try(:table_name)
+    self.names_overriden[model] || model.try(:name)
+  end
+
+  # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
+  def self.name_old_for(model)
+    self.names_old_overriden[model] || model.try(:table_name)
   end
 end