+++ lib/forest_liana/bootstraper.rb
@@ -32,9 +32,15 @@ module ForestLiana
 
     private
 
+    def is_sti_parent_model?(model)
+      return false unless model.try(:table_exists?)
+
+      model.inheritance_column && model.columns.find { |column| column.name == model.inheritance_column }
+    end
+
     def analyze_model?(model)
       model && model.table_exists? && !SchemaUtils.habtm?(model) &&
-        SchemaUtils.model_included?(model) && !SchemaUtils.sti_child?(model)
+        SchemaUtils.model_included?(model)
     end
 
     def fetch_models
@@ -46,6 +52,10 @@ module ForestLiana
         if model.abstract_class?
           model.descendants.each { |submodel| fetch_model(submodel) }
         else
+          if is_sti_parent_model?(model)
+            model.descendants.each { |submodel_sti| fetch_model(submodel_sti) }
+          end
+
           if analyze_model?(model)
             ForestLiana.models << model
           end
@@ -109,11 +119,14 @@ module ForestLiana
         .select { |table_name, models| models.length > 1 }
         .try(:each) do |table_name, models|
           models.each do |model|
-            ForestLiana.names_overriden[model] =
+            unless model.name.deconstantize.blank?
+              ForestLiana.names_overriden[model] = model.name.gsub('::', '__')
+            end
+            # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
+            ForestLiana.names_old_overriden[model] =
               "#{model.name.deconstantize.downcase}__#{model.table_name}"
           end
         end
-
     end
 
     def create_apimap
@@ -206,10 +219,13 @@ module ForestLiana
 
     def setup_intercom_integration(collection_name)
       model_name = ForestLiana.name_for(collection_name.constantize)
+      # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
+      model_name_old = ForestLiana.name_old_for(collection_name.constantize)
       collection_display_name = collection_name.capitalize
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_intercom_conversations",
+        name_old: "#{model_name_old}_intercom_conversations",
         display_name: collection_display_name + ' Conversations',
         icon: 'intercom',
         integration: 'intercom',
@@ -227,6 +243,7 @@ module ForestLiana
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_intercom_attributes",
+        name_old: "#{model_name_old}_intercom_attributes",
         display_name: collection_display_name + ' Attributes',
         icon: 'intercom',
         integration: 'intercom',
@@ -270,10 +287,13 @@ module ForestLiana
     def setup_stripe_integration(collection_name_and_field)
       collection_name = collection_name_and_field.split('.')[0]
       model_name = ForestLiana.name_for(collection_name.constantize)
+      # TODO: Remove once lianas prior to 2.0.0 are not supported anymore.
+      model_name_old = ForestLiana.name_old_for(collection_name.constantize)
       collection_display_name = model_name.capitalize
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_payments",
+        name_old: "#{model_name_old}_stripe_payments",
         display_name: collection_display_name + ' Payments',
         icon: 'stripe',
         integration: 'stripe',
@@ -307,6 +327,7 @@ module ForestLiana
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_invoices",
+        name_old: "#{model_name_old}_stripe_invoices",
         display_name: collection_display_name + ' Invoices',
         icon: 'stripe',
         integration: 'stripe',
@@ -341,6 +362,7 @@ module ForestLiana
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_cards",
+        name_old: "#{model_name_old}_stripe_cards",
         display_name: collection_display_name + ' Cards',
         icon: 'stripe',
         integration: 'stripe',
@@ -376,6 +398,7 @@ module ForestLiana
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_subscriptions",
+        name_old: "#{model_name_old}_stripe_subscriptions",
         display_name: collection_display_name + ' Subscriptions',
         icon: 'stripe',
         integration: 'stripe',
@@ -409,6 +432,7 @@ module ForestLiana
 
       ForestLiana.apimap << ForestLiana::Model::Collection.new({
         name: "#{model_name}_stripe_bank_accounts",
+        name_old: "#{model_name_old}_stripe_bank_accounts",
         display_name: collection_display_name + ' Bank Accounts',
         icon: 'stripe',
         integration: 'stripe',