+++ test/services/forest_liana/schema_adapter_test.rb
@@ -40,7 +40,7 @@ module ForestLiana
       assert schema.fields.include?({
         field: :has_one_field,
         type: 'Number',
-        reference: 'has_one_fields.id',
+        reference: 'HasOneField.id',
         inverseOf: 'belongs_to_field'
       })
     end
@@ -50,7 +50,7 @@ module ForestLiana
       assert schema.fields.include?({
         field: 'belongs_to_field',
         type: 'Number',
-        reference: 'belongs_to_fields.id',
+        reference: 'BelongsToField.id',
         inverseOf: 'has_one_field',
         "is-filterable": true
       })
@@ -61,7 +61,7 @@ module ForestLiana
       assert schema.fields.include?({
         field: 'belongs_to_fields',
         type: ['Number'],
-        reference: 'belongs_to_fields.id',
+        reference: 'BelongsToField.id',
         inverseOf: 'has_many_field',
         "is-filterable": false
       })
@@ -72,7 +72,7 @@ module ForestLiana
       assert schema.fields.include?({
         field: 'foo',
         type: ['Number'],
-        reference: 'belongs_to_fields.id',
+        reference: 'BelongsToField.id',
         inverseOf: 'has_many_class_name_field',
         "is-filterable": false
       })
@@ -83,7 +83,7 @@ module ForestLiana
       assert schema.fields.include?({
         field: :foo,
         type: 'Number',
-        reference: 'has_one_fields.id',
+        reference: 'HasOneField.id',
         inverseOf: 'belongs_to_class_name_field'
       })
     end